Restructure project files
- Created routes package
- Add tablet screens
